requirements.txt
Django==2.0.3
mysqlclient==1.3.12
google-cloud-speech==0.32.1
google-cloud-storage==1.8.0